Install as rclcpp component
